Add yoyo integration

+++ db/db_creator.py
@@ -1,22 +1,31 @@
 # standard libraries
-import logging
+import logging, os
 from typing import Dict, Sequence
 
 # third-party libraries
-from sqlalchemy.engine import Engine
+from sqlalchemy.engine import create_engine
+from yoyo import get_backend, read_migrations
 
 logger = logging.getLogger(__name__)
 
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
+MIGRATIONS_PATH = os.path.join(ROOT_DIR, 'migrations')
 
-# Class(es)
 
 class DBCreator:
 
-    def __init__(self, db_name: str, engine: Engine, table_dicts: Sequence[Dict[str, str]]) -> None:
+    def __init__(self, db_params: Dict[str, str]) -> None:
+        self.db_name = db_params['dbname']
         self.conn = None
-        self.db_name = db_name
-        self.engine = engine
-        self.tables = table_dicts
+        self.conn_str = (
+            'mysql+mysqldb' +
+            f"://{db_params['user']}" +
+            f":{db_params['password']}" +
+            f"@{db_params['host']}" +
+            f":{db_params['port']}" +
+            f"/{db_params['dbname']}?charset=utf8"
+        )
+        self.engine = create_engine(self.conn_str)
 
     def set_up(self) -> None:
         logger.debug('set_up')
@@ -26,34 +35,29 @@ class DBCreator:
         logger.debug('tear_down')
         self.conn.close()
 
-    def drop_tables(self) -> None:
-        logger.debug('drop_tables')
-        self.conn.execute('SET FOREIGN_KEY_CHECKS=0;')
-        table_names = [table['name'] for table in self.tables]
-        drop_statement = f'DROP TABLE IF EXISTS {", ".join(table_names)};'
-        self.conn.execute(drop_statement)
-        self.conn.execute('SET FOREIGN_KEY_CHECKS=1;')
+    def get_table_names(self) -> Sequence[str]:
+        logger.debug('get_table_names')
+        return self.engine.table_names()
+
+    def migrate(self) -> None:
+        logger.debug('migrate')
+        backend = get_backend(self.conn_str)
+        migrations = read_migrations(MIGRATIONS_PATH)
+        with backend.lock():
+            backend.apply_migrations(backend.to_apply(migrations))
 
     def drop_records(self) -> None:
         logger.debug('drop_records')
         self.conn.execute('SET FOREIGN_KEY_CHECKS=0;')
-        for table_dict in self.tables:
-            table_name = table_dict['name']
-            logger.debug(f'Table Name: {table_name}')
-            self.conn.execute(f'DELETE FROM {table_name};')
-            logger.info(f'Dropped records in {table_name} in {self.db_name}')
+        for table_name in self.get_table_names():
+            if 'yoyo' not in table_name:
+                logger.debug(f'Table Name: {table_name}')
+                self.conn.execute(f'DELETE FROM {table_name};')
+                logger.info(f'Dropped records in {table_name} in {self.db_name}')
         self.conn.execute('SET FOREIGN_KEY_CHECKS=1;')
-
-    def create_tables(self) -> None:
-        logger.debug('create_tables')
-        for table_dict in self.tables:
-            table_name = table_dict['name']
-            logger.debug(f'Table Name: {table_name}')
-            self.conn.execute(table_dict['statement'])
-            logger.info(f'Created table {table_name} in {self.db_name}')
 
     def set_up_database(self) -> None:
         self.set_up()
-        self.drop_tables()
-        self.create_tables()
+        self.drop_records()
+        self.migrate()
         self.tear_down()
